Extracted shared ESLint rules + filename shim into root eslint.shared.mjs
no ref

- added eslint.shared.mjs at repo root exporting: correctnessRules (12 shared rules), tsUnusedVarsRule, jsUnusedVarsRule, sortImportsRule, shadeLayeredImportsRule, reactDefaultsOff, reactStrictRules, tailwindRulesV4, tailwindRulesWithConfig(cfg) for v3, mochaRulesOff(ghostPlugin), and localFilenamesPlugin (an ESLint 9-compatible replacement for eslint-plugin-filenames-ts's match-regex)
- refactored 14 React/TS workspace flat configs (apps/activitypub, admin-x-design-system, admin-x-framework, admin-x-settings, announcement-bar, comments-ui, portal, posts, shade, signup-form, sodo-search, stats; ghost/i18n, ghost/parse-email-address) to import from the shared module — each config keeps its workspace-specific extras inline
- replaced inline filenamesMatchRegex shims in ghost/core, ghost/admin, and ghost/i18n with the shared localFilenamesPlugin (the shared version includes ignoreExporting support, previously only in ghost/core)
- ghost/i18n explicitly disables ghost/filenames/match-regex (added by correctnessRules) since this workspace uses the local-filenames variant instead
- net: -450 lines across 16 configs; lint output verified byte-identical to baseline for every workspace touched
