+++ apps/activitypub/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@tryghost/activitypub",
   "type": "module",
-  "version": "3.1.45",
+  "version": "3.1.46",
   "license": "MIT",
   "repository": {
     "type": "git",
