+++ apps/comments-ui/eslint.config.js
@@ -6,32 +6,15 @@ import i18nextPlugin from 'eslint-plugin-i18next';
 import tailwindcssPlugin from 'eslint-plugin-tailwindcss';
 import tseslint from 'typescript-eslint';
 
-const tailwindConfig = `${import.meta.dirname}/tailwind.config.js`;
+import {
+    correctnessRules,
+    reactDefaultsOff,
+    sortImportsRule,
+    tailwindRulesWithConfig,
+    tsUnusedVarsRule
+} from '../../eslint.shared.mjs';
 
-const ghostRules = {
-    curly: 'error',
-    camelcase: ['error', {properties: 'never'}],
-    'dot-notation': 'error',
-    eqeqeq: ['error', 'always'],
-    'no-plusplus': ['error', {allowForLoopAfterthoughts: true}],
-    'no-eval': 'error',
-    'no-useless-call': 'error',
-    'no-console': 'error',
-    'no-shadow': 'error',
-    'array-callback-return': 'error',
-    'no-constructor-return': 'error',
-    'no-promise-executor-return': 'error',
-    'no-unused-vars': 'off',
-    '@typescript-eslint/no-unused-vars': ['error', {
-        args: 'after-used',
-        argsIgnorePattern: '^_',
-        caughtErrors: 'none'
-    }],
-    'ghost/filenames/match-regex': ['error', '^[a-z0-9.-]+$', false],
-    'ghost/sort-imports-es6-autofix/sort-imports-es6': ['error', {
-        memberSyntaxSortOrder: ['none', 'all', 'single', 'multiple']
-    }]
-};
+const tailwindConfig = `${import.meta.dirname}/tailwind.config.js`;
 
 const reactFlat = reactPlugin.configs.flat.recommended;
 const i18nextFlat = i18nextPlugin.configs['flat/recommended'];
@@ -65,11 +48,10 @@ export default tseslint.config(
             ...js.configs.recommended.rules,
             ...reactFlat.rules,
             ...i18nextFlat.rules,
-            ...ghostRules,
-            '@typescript-eslint/no-inferrable-types': 'off',
-            '@typescript-eslint/no-explicit-any': 'warn',
-            'react/react-in-jsx-scope': 'off',
-            'react/prop-types': 'off',
+            ...correctnessRules,
+            ...tsUnusedVarsRule,
+            ...sortImportsRule,
+            ...reactDefaultsOff,
             'react/jsx-sort-props': ['error', {
                 reservedFirst: true,
                 callbacksLast: true,
@@ -78,13 +60,9 @@ export default tseslint.config(
             }],
             'react/button-has-type': 'error',
             'react/no-array-index-key': 'error',
-            'tailwindcss/classnames-order': ['error', {config: tailwindConfig}],
-            'tailwindcss/enforces-negative-arbitrary-values': ['warn', {config: tailwindConfig}],
-            'tailwindcss/enforces-shorthand': ['warn', {config: tailwindConfig}],
-            'tailwindcss/migration-from-tailwind-2': ['warn', {config: tailwindConfig}],
-            'tailwindcss/no-arbitrary-value': 'off',
-            'tailwindcss/no-custom-classname': 'off',
-            'tailwindcss/no-contradicting-classname': ['error', {config: tailwindConfig}],
+            ...tailwindRulesWithConfig(tailwindConfig),
+            '@typescript-eslint/no-inferrable-types': 'off',
+            '@typescript-eslint/no-explicit-any': 'warn',
             'no-undef': 'off'
         }
     }
