+++ ghost/i18n/eslint.config.mjs
@@ -1,50 +1,15 @@
-import path from 'node:path';
 import js from '@eslint/js';
 import globals from 'globals';
 import ghostPlugin from 'eslint-plugin-ghost';
 
-// eslint-plugin-filenames-ts@1.3.2's match-regex rule still calls
-// context.getScope(), which ESLint 9 removed. Replace it with a minimal
-// equivalent that only does the filename check (no isExporting / isExportingClass).
-const filenamesMatchRegex = {
-    meta: {
-        type: 'problem',
-        schema: [{type: 'string'}, {type: ['boolean', 'null']}, {type: ['boolean', 'null']}]
-    },
-    create(context) {
-        const pattern = new RegExp(context.options[0]);
-        return {
-            Program(node) {
-                const filename = path.parse(context.filename).name;
-                if (!pattern.test(filename)) {
-                    context.report({
-                        node,
-                        message: `Filename '${filename}' does not match the naming convention.`
-                    });
-                }
-            }
-        };
-    }
-};
+import {
+    correctnessRules,
+    jsUnusedVarsRule,
+    localFilenamesPlugin,
+    mochaRulesOff
+} from '../../eslint.shared.mjs';
 
-const localFilenamesPlugin = {
-    rules: {'match-regex': filenamesMatchRegex}
-};
-
-const ghostRules = {
-    curly: 'error',
-    camelcase: ['error', {properties: 'never'}],
-    'dot-notation': 'error',
-    eqeqeq: ['error', 'always'],
-    'no-plusplus': ['error', {allowForLoopAfterthoughts: true}],
-    'no-eval': 'error',
-    'no-useless-call': 'error',
-    'no-console': 'error',
-    'no-shadow': 'error',
-    'array-callback-return': 'error',
-    'no-constructor-return': 'error',
-    'no-promise-executor-return': 'error',
-    'no-unused-vars': ['error', {caughtErrors: 'none'}],
+const ghostI18nExtras = {
     'no-var': 'warn',
     'one-var': ['warn', 'never'],
     'ghost/node/no-restricted-require': ['warn', [
@@ -56,14 +21,11 @@ const ghostRules = {
     'ghost/ghost-custom/no-native-error': 'error',
     'ghost/ghost-custom/ghost-error-usage': 'error',
     'ghost/ghost-custom/ghost-tpl-usage': 'error',
+    // This workspace uses the local-filenames variant of the rule; turn off
+    // the eslint-plugin-ghost one that correctnessRules enables.
+    'ghost/filenames/match-regex': 'off',
     'local-filenames/match-regex': ['error', '^[a-z0-9.-]+$', false]
 };
-
-const mochaRulesOff = Object.fromEntries(
-    Object.keys(ghostPlugin.rules || {})
-        .filter(rule => rule.startsWith('mocha/'))
-        .map(rule => [`ghost/${rule}`, 'off'])
-);
 
 export default [
     {
@@ -83,7 +45,9 @@ export default [
         },
         rules: {
             ...js.configs.recommended.rules,
-            ...ghostRules
+            ...correctnessRules,
+            ...jsUnusedVarsRule,
+            ...ghostI18nExtras
         }
     },
     {
@@ -114,8 +78,10 @@ export default [
         },
         rules: {
             ...js.configs.recommended.rules,
-            ...ghostRules,
-            ...mochaRulesOff,
+            ...correctnessRules,
+            ...jsUnusedVarsRule,
+            ...ghostI18nExtras,
+            ...mochaRulesOff(ghostPlugin),
             'ghost/ghost-custom/node-assert-strict': 'error'
         }
     }
