+++ ghost/parse-email-address/eslint.config.mjs
@@ -3,38 +3,19 @@ import globals from 'globals';
 import ghostPlugin from 'eslint-plugin-ghost';
 import tseslint from 'typescript-eslint';
 
-const ghostRules = {
-    curly: 'error',
-    camelcase: ['error', {properties: 'never'}],
-    'dot-notation': 'error',
-    eqeqeq: ['error', 'always'],
-    'no-plusplus': ['error', {allowForLoopAfterthoughts: true}],
-    'no-eval': 'error',
-    'no-useless-call': 'error',
-    'no-console': 'error',
-    'no-shadow': 'error',
-    'array-callback-return': 'error',
-    'no-constructor-return': 'error',
-    'no-promise-executor-return': 'error',
-    'no-unused-vars': 'off',
-    '@typescript-eslint/no-unused-vars': ['error', {
-        args: 'after-used',
-        argsIgnorePattern: '^_',
-        caughtErrors: 'none'
-    }],
+import {
+    correctnessRules,
+    mochaRulesOff,
+    tsUnusedVarsRule
+} from '../../eslint.shared.mjs';
+
+const ghostParseEmailExtras = {
     'no-var': 'warn',
     'one-var': ['warn', 'never'],
     'ghost/ghost-custom/no-native-error': 'error',
     'ghost/ghost-custom/ghost-error-usage': 'error',
-    'ghost/ghost-custom/ghost-tpl-usage': 'error',
-    'ghost/filenames/match-regex': ['error', '^[a-z0-9.-]+$', false]
+    'ghost/ghost-custom/ghost-tpl-usage': 'error'
 };
-
-const mochaRulesOff = Object.fromEntries(
-    Object.keys(ghostPlugin.rules || {})
-        .filter(rule => rule.startsWith('mocha/'))
-        .map(rule => [`ghost/${rule}`, 'off'])
-);
 
 export default tseslint.config(
     {
@@ -53,7 +34,9 @@ export default tseslint.config(
         },
         rules: {
             ...js.configs.recommended.rules,
-            ...ghostRules,
+            ...correctnessRules,
+            ...tsUnusedVarsRule,
+            ...ghostParseEmailExtras,
             'no-undef': 'off'
         }
     },
@@ -77,8 +60,10 @@ export default tseslint.config(
         },
         rules: {
             ...js.configs.recommended.rules,
-            ...ghostRules,
-            ...mochaRulesOff,
+            ...correctnessRules,
+            ...tsUnusedVarsRule,
+            ...ghostParseEmailExtras,
+            ...mochaRulesOff(ghostPlugin),
             'no-undef': 'off'
         }
     }
